Removed p.a.widgets moved DeprecationWarning
