+++ news/690.bugfix
@@ -1,0 +1,1 @@
+Fixed p.a.widgets `moved` DeprecationWarning [jensens]
